A-Packages: Separate files into different class files and packages
