+++ src/main/java/jackbit/task/Todo.java
@@ -1,0 +1,13 @@
+package jackbit.task;
+
+public class Todo extends Task {
+
+    public Todo(String name) {
+        super(name);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
